% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Element_Data.R
\name{add_element_data}
\alias{add_element_data}
\title{Add ionic radius and chondrite and mantle values, Z and Mass}
\usage{
add_element_data(dat)
}
\arguments{
\item{dat}{Long data REE format}
}
\value{
A data frame
}
\description{
This is a helper function to work with Element_norm() and Element_denorm().
Add Ionic Radius to data and chondrite values. For now, only supports 3+ in
eight-fold
coordination for REE, Zr and Y.Values are from Shannon(1976),
McDonough and Sun (1995) and Palme and O'Neill (2014).
}
