% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_ID.R
\name{Add_ID}
\alias{Add_ID}
\title{Add_ID}
\usage{
Add_ID(dat, ID = "rowid", ...)
}
\arguments{
\item{dat}{a tibble or a dataframe}

\item{ID}{Name of column to use for rownames.}

\item{...}{Other parameters passed onto the \code{tibble::rowid_to_column()} function}
}
\value{
a data frame
}
\description{
Add an unique ID per observation and checks that is not overwriting an existing column.
}
\examples{

Element_Data \%>\% Add_ID()

}
