% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Norm_Values.R
\name{Add_NormValues}
\alias{Add_NormValues}
\title{Add Chondrite or Mantle values for normalization.}
\usage{
Add_NormValues(dat, method = PalmeOneill2014CI)
}
\arguments{
\item{dat}{Dataframe or tibble.}

\item{method}{Values to normalize: an option from: PalmeOneill2014CI, Oneill2014Mantle, McDonough1995CI.
load}
}
\value{
a data frame or tibble
}
\description{
This is a helper function to work with Element_norm() and Element_denorm(). Takes long pivoted data to match element name and add normalizing values from the Element_data dataset.
}
\details{
This is a test to see what happens in the following paragrapfs
}
\examples{

Data <- testing_data \%>\%  CleanColnames('Zr', 'ppm') \%>\%  Add_ID()

Data \%>\%
dplyr::select(rowid, La, Ce) \%>\%
tidyr::pivot_longer(cols = -rowid, names_to = 'Element_name')  \%>\%
Add_NormValues()

}
\seealso{
\code{\link{Element_norm()}}

Other Add Functions: 
\code{\link{Add_IonicRadii}()}
}
\concept{Add Functions}
