% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_parameters.R, R/calc_anomalies.R
\name{Add_parameters}
\alias{Add_parameters}
\title{Title}
\usage{
Add_parameters(data, preffix = NULL, suffix = NULL)

Add_parameters(data, preffix = NULL, suffix = NULL)
}
\arguments{
\item{suffix}{}
}
\value{



}
\description{
Title

Title
}
