% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColnames.R
\name{CleanColnames}
\alias{CleanColnames}
\title{Clean variable names that have prefixes or suffixes}
\usage{
CleanColnames(dat, prefix = NULL, suffix = NULL)
}
\arguments{
\item{dat}{a dataframe}

\item{prefix}{A character of lengt 1}

\item{suffix}{A character of lengt 1}
}
\value{
A dataframe
}
\description{
This is a helper function
}
