% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Element_Data.R
\name{create_REE_Data}
\alias{create_REE_Data}
\title{Create Element Data}
\usage{
create_REE_Data()
}
\value{
tibble
}
\description{
Auxiliar Function that create data. Format is prepared for easy update. Contained within the package.
Explanation is in Element_Data dataset.
}
