% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoItALL.R
\name{DoItALL}
\alias{DoItALL}
\title{Title}
\usage{
DoItALL(data, preffix = NULL, suffix = NULL, method = PalmeOneill2014CI)
}
\arguments{
\item{data}{a dataframe}

\item{preffix}{A character of lengt 1}

\item{suffix}{A character of lengt 1}

\item{method}{an option from: PalmeOneill2014CI, Oneill2014Mantle, McDonough1995CI}
}
\description{
This is a wrapper for data \%>\% Model_REE() \%>\% Imputate_REE \%>\% Add_parameters()
}
