% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Element_denorm.R
\name{element_denorm}
\alias{element_denorm}
\title{Denormalize chrodrite Normalize to ppm}
\usage{
element_denorm(dat, method = PalmeOneill2014CI)
}
\arguments{
\item{dat}{A dataframe}

\item{method}{an option from: 'PalmeOneill2014CI', 'Oneill2014Mantle', 'McDonough1995CI'}
}
\value{
A dataframe
}
\description{
Denormalize chrodrite Normalize to ppm
}
