% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputate_REE.R
\name{Imputate_REE}
\alias{Imputate_REE}
\title{Impute Rare earth elememts}
\usage{
Imputate_REE(data, preffix = NULL, suffix = NULL, rsquared = 0.9)
}
\arguments{
\item{rsquared}{A numerical value between 0 and 1. Tolerance to missfiting models.}
}
\value{
A dataframe
}
\description{
Imputes missing REE after modelling. Expect the output of Model_REE() function
}
\details{
By default, exclude models with Rsqured lower than 0.9.
}
