% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REE_Elements.R
\docType{data}
\name{REE_Elements}
\alias{REE_Elements}
\title{Rare earth element list}
\format{
Rare earth element list
}
\usage{
REE_Elements
}
\description{
A string vector containing the elemental symbols for REE.
}
\keyword{datasets}
