% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REE_Normalizer.R
\name{REE_norm}
\alias{REE_norm}
\title{Title}
\usage{
REE_norm(
  dat,
  return = "rect",
  method = Oneill2014CI,
  prefix = "Zr",
  sufix = "ppm",
  Element_list = REE_plus_Y_Elements,
  ...
)
}
\arguments{
\item{dat}{a datafram}

\item{return}{a characther from: "rect" for a wide data return,"raw" for a long data return,"append" to append the results to the input data}

\item{method}{an opction from: Oneill2014CI, Oneill2014Mantle, McDonough1995CI}

\item{prefix}{a character: by defaulkt it is 'Zr'}

\item{sufix}{a character: by default it expect to be the unit, usually in 'ppm'}

\item{Element_list}{a character vector: indicating the elements that should be normalized}

\item{...}{Additional arguments passed on to functions.}
}
\value{
a data frame
}
\description{
Title
}
\examples{
testing_data \%>\%  REE_norm()

}
