% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REE_plus_Y_Elements.R
\docType{data}
\name{REE_plus_Y_Elements}
\alias{REE_plus_Y_Elements}
\title{Rare earth element list}
\format{
Rare earth element + Y list
}
\usage{
REE_plus_Y_Elements
}
\description{
A string vector containing the elemental symbols for REE and Y.
}
\keyword{datasets}
