% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_ID.R
\name{add_ID}
\alias{add_ID}
\title{Add_ID}
\usage{
add_ID(dat, ID = "rowid")
}
\arguments{
\item{dat}{a tibble or a dataframe}

\item{ID}{Name of column to use for rownames. 'rowid' is used if none is specified.
er parameters passed onto the \code{tibble::rowid_to_column()} function}
}
\value{
a data frame
}
\description{
Add an unique ID per observation and checks that is not overwriting an existing column. If the column already exist, it will take no action.
This is a wrapper of tibble::rowid_to_column() that checks that not columns is overwritten.
}
