% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Ionic_Radii.R
\name{Add_IonicRadii}
\alias{Add_IonicRadii}
\title{Add Chondrite or Mantle values for normalization.}
\usage{
Add_IonicRadii(dat, method = ShannonRadiiVIII_Coord_3plus)
}
\arguments{
\item{dat}{a dataframe or tibble.}

\item{method}{Ionic Radii from shannon1976. Not.}
}
\value{
a data frame or tibble
}
\description{
This is a helper function to work with Element_norm() and Element_denorm(). Takes long pivoted data to match element name and add normalizing values from the Element_data dataset.
}
\details{
This is a test to see what happens in the following paragrapfs
}
\examples{

testing_data \%>\%
CleanColnames('Zr', 'ppm') \%>\%  Add_ID() \%>\%
dplyr::select(rowid, La, Ce) \%>\%
tidyr::pivot_longer(cols = -rowid, names_to = 'Element_name')  \%>\%
Add_IonicRadii()

}
\seealso{
\code{\link{Element_norm}}

Other Add Functions: 
\code{\link{Add_NormValues}()}
}
\concept{Add Functions}
