% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Ionic_Radii.R
\name{add_IonicRadii}
\alias{add_IonicRadii}
\title{Add Chondrite or Mantle values for normalization.}
\usage{
add_IonicRadii(dat, method = ShannonRadiiVIII_Coord_3plus)
}
\arguments{
\item{dat}{a dataframe or tibble.}

\item{method}{Ionic Radii from Shannon, 1976}
}
\value{
a data frame or tibble
}
\description{
This is a helper function to work with Element_norm() and Element_denorm(). Takes long pivoted data to match element name and add normalizing values from the Element_data dataset.
}
