% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Norm_Values.R
\name{Add_NormValues}
\alias{Add_NormValues}
\title{Add Chondrite or Mantle values for normalization.}
\usage{
Add_NormValues(dat, method = PalmeOneill2014CI)
}
\arguments{
\item{dat}{Dataframe or tibble.}

\item{method}{Values to normalize: an option from: PalmeOneill2014CI, Oneill2014Mantle, McDonough1995CI.}
}
\value{
a data frame or tibble
}
\description{
This is a helper function to work with Element_norm() and Element_denorm(). Takes long pivoted data to match element name and add normalizing values from the Element_data dataset.
}
\details{
This is a test to see what happens in the following paragrapfs
}
\seealso{
\code{\link{Element_norm()}}
}
\concept{Add Functions}
