% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Norm_Values.R
\name{add_NormValues}
\alias{add_NormValues}
\title{Add Chondrite or Mantle values for normalization.}
\usage{
add_NormValues(dat, chondrite = PalmeOneill2014CI)
}
\arguments{
\item{dat}{Dataframe or tibble.
doc}

\item{chondrite}{PalmeOneill2014CI, Oneill2014Mantle, McDonough1995CI}
}
\value{
a data frame or tibble
}
\description{
This is a helper function to work with Element_norm() and Element_denorm(). Takes long pivoted data to match element name and add normalizing values from the Element_data dataset.
}
