% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColnames.R
\name{CleanColnames}
\alias{CleanColnames}
\title{Clean variable names that have prefixes or suffixes}
\usage{
CleanColnames(dat, prefix = NULL, suffix = NULL)
}
\arguments{
\item{dat}{a data frame}

\item{prefix}{A character of length 1}

\item{suffix}{A character of length 1}
}
\value{
A data frame
}
\description{
This is a helper function
}
