% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColnames.R
\name{CleanColnames}
\alias{CleanColnames}
\title{Title}
\usage{
CleanColnames(dat, preffix = NULL, suffix = NULL)
}
\arguments{
\item{preffix}{A character of lengt 1}

\item{suffix}{A character of lengt 1}
}
\value{
A dataframe
}
\description{
Title
}
\examples{

testing_data \%>\% CleanColnames(preffix = 'Zr', suffix = 'ppm')

}
