% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoItALL.R
\name{DoItALL}
\alias{DoItALL}
\title{Title}
\usage{
DoItALL(data, preffix = NULL, suffix = NULL, method = PalmeOneill2014CI)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
This is a wrapper for data \%>\% Model_REE() \%>\% Imputate_REE \%>\% Add_parameters()
}
