% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Element_norm.R
\name{Element_norm}
\alias{Element_norm}
\title{Calculate normalized values for a list of elements}
\usage{
Element_norm(
  dat,
  return = "rect",
  method = Oneill2014CI,
  prefix = "Zr",
  sufix = "ppm",
  Element_list = REE_plus_Y_Elements,
  ID = "rowid"
)
}
\arguments{
\item{dat}{a dataframe}

\item{return}{a characther from: "rect" for a wide data return,"raw" for a long data return,"append" to append the results to the input data}

\item{method}{an opction from: Oneill2014CI, Oneill2014Mantle, McDonough1995CI}

\item{prefix}{a character: by defaulkt it is 'Zr'}

\item{sufix}{a character: by default it expect to be the unit, usually in 'ppm'}

\item{Element_list}{a character vector: indicating the elements that should be normalized}

\item{ID}{Name of column to use for rownames. Is name is alredy in use, an error message will ask for a new name}
}
\value{
a data frame
}
\description{
Element norm normalize values according to published values for the Primitive mantle and chondrites. By defect, it uses the values from Palme and O'Neill (2014). By default, REE + Y list is provided
}
\examples{
testing_data \%>\%  Element_norm()

}
