% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_REE.R
\name{Model_REE}
\alias{Model_REE}
\title{Model REE + Y contents using an empirical method based on the lattice strain theory}
\usage{
Model_REE(
  dat,
  r0 = 0.84,
  exclude = c("La", "Ce", "Eu", "Y"),
  preffix = NULL,
  suffix = NULL,
  method = PalmeOneill2014CI,
  Y_correction_fact = 1.29
)
}
\arguments{
\item{dat}{A data frame}

\item{r0}{A number: ionic radii of the lattice site r0}

\item{exclude}{a string: vector including elements that should be ommited from modelling. La, Ce and Eu are the default. Ce and Eu should be always included}

\item{preffix}{A character of lengt 1}

\item{suffix}{A character of lengt 1}

\item{method}{an option from: PalmeOneill2014CI, Oneill2014Mantle, McDonough1995CI}

\item{Y_correction_fact}{a number: correction factor for underestimated Y. 1.29 by default.}
}
\value{
a dataframe
}
\description{
Model REE will make a linear regression between the REE (+Y) and the relationship of the ideal Ionic Radii in the lattice site (r0) and the ionic radii of the element that use that space (ri) according to the relationship : (ri/3 + r0/6)(ri-r0)^2`
}
\details{
For details in the lattice strain theory, see Blundy and Wood 1994.
}
\examples{

testing_data \%>\%  Model_REE(preffix = 'Zr', suffix = 'ppm')
}
